% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_variables_describe.R
\name{choose_variables_describe}
\alias{choose_variables_describe}
\title{Read variables from excel file, process and save}
\usage{
choose_variables_describe(inxlfile, vmapr = NULL, vmapspss = NULL,
  vmapstata = NULL, dr = NULL, ds = NULL, dt = NULL)
}
\arguments{
\item{inxlfile}{Excel input file}

\item{vmapr}{Map of R variables}

\item{vmapspss}{Map of SPSS variables}

\item{vmapstata}{Map of STATA variables}

\item{dr}{File of R variables}

\item{ds}{File of SPSS variables}

\item{dt}{File of STATA variables}
}
\description{
Read user specified variables with instructions, create piecharts, barplots and save frequencies of answers
}
\examples{
#choose_variables_describe(inxlfile,vmapr,vmapspss,vmapstata,dr,ds,dt)
}
