% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_analysis.R
\name{country_analysis}
\alias{country_analysis}
\title{Create country mixtures for archetypes and indices HH, NH, MI}
\usage{
country_analysis(wavedatain, varnames, archetypesin, varcat = "Schwartz",
  exclude = FALSE)
}
\arguments{
\item{wavedatain}{The rds file with columns the World Values to be used}

\item{varnames}{R The vector of names of values to be used}

\item{archetypesin}{The rds file with columns the final archetype found}

\item{varcat}{The character name of values category (default='Schwartz')}

\item{exclude}{Logical entry set to FALSE}
}
\description{
Read processed data frame, read variables, archetype data frame, variable category and 
create country mixtures for archetypes plus indices HH, NH, MI
}
\examples{
#country_analysis(wavedatain,varnames,archetypesin)
}
