% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalues-package.R
\docType{package}
\name{cvalues-package}
\alias{cvalues}
\alias{cvalues-package}
\title{cvalues: Read, Analyze and Visualize World Values Surveys Data}
\description{
Read World Values Survey Data in R, SPSS, STATA Format and Visualize Them.
}
\author{
\strong{Maintainer}: Demetris Christopoulos \email{dem.christop@gmail.com}

Other contributors:
\itemize{
  \item David Midgley \email{david.midgley@insead.edu} [contributor, copyright holder]
  \item INSEAD Fontainebleau France [funder, copyright holder]
}

}
\keyword{internal}
