% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Final_W6_Archetypes}
\alias{Final_W6_Archetypes}
\title{The final data set for archetypal mixture for the 5 archetypes created in WVS wave 6}
\format{A data frame with 5 rows and 11 variables:
\describe{
\item{B_Archetype}{Alpha label of archetype} 
\item{Universalism}{score of Schwartz value} 
\item{Benevolence}{score of Schwartz value} 
\item{Tradition}{score of Schwartz value} 
\item{Conformity}{score of Schwartz value} 
\item{Security}{score of Schwartz value} 
\item{Power}{score of Schwartz value} 
\item{Achievement}{score of Schwartz value} 
\item{Hedonism}{score of Schwartz value} 
\item{Stimulation}{score of Schwartz value} 
\item{Self-direction}{score of Schwartz value} 
}}
\usage{
Final_W6_Archetypes
}
\description{
It contains the components of archetypes for the 10 Schwartz values
}
\keyword{datasets}
