% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_variables.R
\name{impute_variables}
\alias{impute_variables}
\title{Impute requested R variables using package 'mice' in parallel mode}
\usage{
impute_variables(inxlfile, df, ncores = NULL)
}
\arguments{
\item{inxlfile}{Excel file of variables to be imputed}

\item{df}{R rds file of data with missing values to be imputed}

\item{ncores}{the number of logical cores that should be used (apply for multicore computers)}
}
\description{
Read excel file of description for imputing variables, read missing values data frame, impute and save results
}
\examples{
#impute_variables_R(inxlfile,df)
}
