% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_country_mixtures.R
\name{mds_country_mixtures}
\alias{mds_country_mixtures}
\title{Create MDS configuration for countries}
\usage{
mds_country_mixtures(n_arches, ndim = 3, countrydata, exclude = FALSE,
  dmetric = "standard", usebest = FALSE, perm = TRUE, jack = TRUE,
  use_indices = TRUE)
}
\arguments{
\item{n_arches}{The integer number of archetypes used}

\item{ndim}{The integer dimension of MDS configuration}

\item{countrydata}{The excel file with country data to be used as input}

\item{exclude}{Logical entry set to FALSE}

\item{dmetric}{The distance metric to be used}

\item{usebest}{Logical entry set to FALSE if Torgerson will be used}

\item{perm}{Logical entry set to TRUE if permutation will be performed}

\item{jack}{Logical entry set to TRUE if jackknife will be performed}

\item{use_indices}{Logical entry set to TRUE if country indices will be used}
}
\description{
Read country data and parameters in order ot create MDS configuration
}
\examples{
#mds_country_mixtures(n_arches,ndim,countrydata)
}
