% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_visual.R
\name{mds_visual}
\alias{mds_visual}
\title{Visualize several MDS variables}
\usage{
mds_visual(mdsconfigsin, mdsregressin, countrymixin, a_names, anamesorder,
  countrydata, plotrgl = TRUE, plot3d = TRUE, plotrgl3d = TRUE,
  plotworld3d = TRUE, plotworld2d = TRUE, plotgoogle = TRUE,
  plotnatearth = TRUE, showiso2 = TRUE, showiso3 = FALSE,
  plotcrop = FALSE, trimplot = TRUE, maxradius = 1e+05,
  col.ellipse = "lightsteelblue", col.ellipse.special = "lightcoral",
  usepositive = TRUE, brewerpal = "RdPu", sagapal = 1,
  col.bubble = "#FF5A0088")
}
\arguments{
\item{mdsconfigsin}{The excel file of MDS configurations for all countries}

\item{mdsregressin}{The excel file of biplot regressions for all countries}

\item{countrymixin}{The excel file with country mixtures}

\item{a_names}{The character name of initially created archetype names}

\item{anamesorder}{The character name of finally chosen name order for archetypes}

\item{countrydata}{The excel file with country data to be used as input}

\item{plotrgl}{Logical entry set to TRUE if rgl plots will be created}

\item{plot3d}{Logical entry set to TRUE if plot3D plots will be created}

\item{plotrgl3d}{Logical entry set to TRUE if plotrgl3d plots will be created}

\item{plotworld3d}{Logical entry set to TRUE if rotatable 3d projected world map plots will be created}

\item{plotworld2d}{Logical entry set to TRUE if 2d world maps will be created}

\item{plotgoogle}{Logical entry set to TRUE if Google maps will be created}

\item{plotnatearth}{Logical entry set to TRUE if 'naturalearth' maps will be created}

\item{showiso2}{Logical entry set to TRUE if ISOAlpha2 labels will be used}

\item{showiso3}{Logical entry set to TRUE if ISOAlpha3 labels will be used}

\item{plotcrop}{Logical entry set to TRUE if knitr::plot_crop command will be used for pdf's}

\item{trimplot}{Logical entry set to TRUE if magick::trimplot command will be used for png's}

\item{maxradius}{Numeric entry for the radius of pies to be plotted in Google Maps (default=100000m)}

\item{col.ellipse}{The color of the inner ellipsoid of concentration}

\item{col.ellipse.special}{The color of the outer ellipsoid of concentration}

\item{usepositive}{Logical entry set to TRUE in rotatable 3d projected world map plots will be created if vertical lines are only above globe}

\item{brewerpal}{The Brewer pallette to be used for 2d maps}

\item{sagapal}{The SAGA pallette to be used for Google Maps}

\item{col.bubble}{The color for bubble plots}
}
\description{
Read configuration inputs and create 3d plots using packages rgl, plot3D, plotrgl3d
Create 2d maps for all variables
Create Google Maps and Google Earth Maps for specific variables
}
\examples{
\dontrun{
#mds_visual(mdsconfigsin,mdsregressin,countrymixin,a_names,anamesorder,
#countrydata,showiso2=TRUE,showiso3=FALSE,plotcrop=TRUE,trimplot=TRUE,
#plotrgl=T,plot3d=T,plotrgl3d=T,plotworld3d=T,plotworld2d=T,plotgoogle=T,plotnatearth=T,
#col.ellipse='lightsteelblue',col.ellipse.special='lightcoral',usepositive=TRUE,
#brewerpal='RdPu',sagapal=1,col.bubble='#FF5A0088',maxradius=1e5)
}
}
