% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_create_map.R
\name{read_data_create_map}
\alias{read_data_create_map}
\title{Read data and create map of variables}
\usage{
read_data_create_map(rin = NULL, spssin = NULL, statain = NULL)
}
\arguments{
\item{rin}{Rdata file}

\item{spssin}{SPSS sav file}

\item{statain}{STATA dta file}
}
\description{
Read World Values Survey data in R, SPSS and STATA format and create variable map
}
\examples{
#Supposed that you have downloaded from
# http://www.worldvaluessurvey.org/WVSDocumentationWV6.jsp next raw data files:
#"WV6_Data_R_v_2016_01_01.rdata","WV6_Data_spss_v_2016_01_01.sav","WV6_Stata_v_2016_01_01.dta"
#read_data_create_map(rin="WV6_Data_R_v_2016_01_01.rdata",
#spssin = "WV6_Data_spss_v_2016_01_01.sav",statain="WV6_Stata_v_2016_01_01.dta")
#Or for just the R data use next:
#read_data_create_map(rin="WV6_Data_R_v_2016_01_01.rdata")
}
