% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_variables.R
\name{reverse_variables}
\alias{reverse_variables}
\title{Reverse requested R variables}
\usage{
reverse_variables(inxlfile, dr, validansR)
}
\arguments{
\item{inxlfile}{Excel file of variables}

\item{dr}{R rds file of data}

\item{validansR}{Valid answers file}
}
\description{
Read excel file of description for reversing variables
}
\examples{
#reverse_variables(inxlfile,dr,validansR=validr)
}
