% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W6}
\alias{W6}
\title{The complete cases data set of World Value Survey Wave 6
after reversing Scwartz values to the scale 
"Not at all like me"->1
...
"Very much like me"->6}
\format{A data frame with 83526 rows and 14 variables:
\describe{
\item{id}{identity stamp for each respondent} 
\item{desc_gender}{gender of respondent: 0=male, 1=female} 
\item{desc_country}{country numeric code} 
\item{cyname}{short country name} 
\item{Universalism}{name of Schwartz value} 
\item{Benevolence}{name of Schwartz value} 
\item{Tradition}{name of Schwartz value} 
\item{Conformity}{name of Schwartz value} 
\item{Security}{name of Schwartz value} 
\item{Power}{name of Schwartz value} 
\item{Achievement}{name of Schwartz value} 
\item{Hedonism}{name of Schwartz value} 
\item{Stimulation}{name of Schwartz value} 
\item{Self-direction}{name of Schwartz value} 
}}
\usage{
W6
}
\description{
The complete cases data set of World Value Survey Wave 6
after reversing Scwartz values to the scale 
"Not at all like me"->1
...
"Very much like me"->6
}
\keyword{datasets}
